Move dependency export scan behind dependency_export_adapter
dependency_artifact no longer imports stdlib ast. Residual dual-body export
selection lives in dependency_export_adapter (retirement: typed SourceFile
export analysis). Whole-kit construction_side_door_law R=0 with sole_path=0.
